pyproject.toml
[project]
name = "social-media-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[all]>=0.128.1",
    "psycopg2-binary>=2.9.11",
    "sqlalchemy>=2.0.46",
]
